Make setup scripts work on ubuntu

+++ mysetup.sh
@@ -1,17 +1,30 @@
 #!/bin/bash
 
+set -e
+
 create_dot_vim() {
-  mkdir ~/.vim
-  ln -s ~/dotfiles/configfiles/UltiSnips ~/.vim/UltiSnips
+  if [ ! -d ~/.vim ]; then
+    echo "creating .vim file"
+    mkdir -p ~/.vim/backup
+  fi
+  if [ ! -e ~/.vim/Ultisnips ]; then
+    ln -s ~/dotfiles/configfiles/UltiSnips ~/.vim/UltiSnips
+  fi
+  if [ -e ~/.config/nvim ]; then
+    rm -rf ~/.config/nvim
+  fi
+  ln -s ~/.vim ~/.config/nvim
 }
 
 copy_dotfiles() {
   echo "copying dotfiles into user directory"
-  cd configfiles
+  cd ~/dotfiles/configfiles
   for file in *; do
-    echo "copying $file"
-    rm ~/."$file"
-    ln -s ~/dotfiles/configfiles/"$file" ~/."$file"
+    if [ -f ~/dotfiles/configfiles/"$file" ]; then
+      [ -f ~/."$file" ] && mv ~/."$file"{,-backup}
+      echo "copying $file"
+      ln -s ~/dotfiles/configfiles/"$file" ~/."$file"
+    fi
   done
 }
 
@@ -29,13 +42,6 @@ vim_swap_and_backup_dirs() {
 vim_install_plug() {
   curl -fLo ~/.vim/autoload/plug.vim --create-dirs \
     https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
-}
-
-install_oh_my_zsh() {
-  git clone git://github.com/robbyrussell/oh-my-zsh.git ~/.oh-my-zsh
-  cp ~/.oh-my-zsh/templates/zshrc.zsh-template ~/.zshrc
-  chsh -s /bin/zsh
-  git clone https://github.com/zsh-users/zsh-completions ~/.oh-my-zsh/custom/plugins/zsh-completions
 }
 
 create_dot_vim
